Optimise size of model structs (#2464)
This commit reduces the sizes of many integer fields, or converts data to enums when appropriate:

- Reduce the size of rate_limit fields
- Reduce the size of channel position
- Reduce size of role position
- Use an enum for `auto_archive_duration`
- `DefaultReaction` -> `ForumEmoji`
- Swap `AfkTimeout` for enum
- Reduce `Sticker::sort_value` to u16
- Reduce size of `min`/`max_values`
- Reduce size of `max_uses` and `max_age`
- Reduce size of `ActivityParty::size`
- Reduce size of `Ready::version`
- Reduce `InputText::min/max_length`
- Reduce `size`/`width`/`height` -- Discord filesize limits mean we won't hit the size limit, and therefore won't hit height/width limits, plus discord rejects 0 sized files so we can use `NonZero*`
- Reduce size of `mention_total_limit`
- Reduce size of `RoleSubscriptionData::total_months_subscribed` -- This gives us ~5000 years of role subscription
- Fix Change sizes

+++ src/model/guild/member.rs
@@ -288,7 +288,7 @@ impl Member {
     /// The "highest role in hierarchy" is defined as the role with the highest position. If two or
     /// more roles have the same highest position, then the role with the lowest ID is the highest.
     #[cfg(feature = "cache")]
-    pub fn highest_role_info(&self, cache: impl AsRef<Cache>) -> Option<(RoleId, u32)> {
+    pub fn highest_role_info(&self, cache: impl AsRef<Cache>) -> Option<(RoleId, u16)> {
         let guild = cache.as_ref().guild(self.guild_id)?;
 
         let mut highest = None;
